Add tests for registration functionality
I've only included the functionality available from the site
itself. (Password reset links work, for example, but are not linked
anywhere on the site)

Note that I had to change the code a little to get it to work with
django-registration-redux. I upgraded that package from 1.1 to 1.3 in a
previous PR in order to get Travis to succeed because version 1.1 did
not have any migrations and 1.2+ did. Unfortunately 1.2 also had an
undocumented breaking change to its API (changing the signature of a
method).