Load most JS from CDN; remove unused template
Changed all the script tags that load standard
libraries to load the same versions from the
cdnjs CDN, which is hosted on Cloudflare.

CSS is a different problem because we import the
CSS library (bootstrap) when compiling our custom
stylesheet and then serve the result of the compilation.

I also deleted open_debates_layout.html, which has
some more script tags but does not seem to be referenced
anywhere.  If it's needed for some reason, I can put
it back.

Is there any reason to keep installing the JS-only
libraries using bower etc.?