Redo headline implementation and fix related behavior in recent activity, merges, etc.
Store user submitted followup text in 'followup' field
instead of 'idea' field, and auto-generate 'idea' field by concatenating headline and
followup text.  This way, the front end can display the two fields as separate whenever
it wants to, and the backend can always refer to the joint 'idea' field when it needs
access to the full text of the question (e.g. for search, merged keywords, etc)

Added now-passing tests for related behaviors:
* When searching questions, headlines contribute to search results.
* When displaying recent submissions, headlines are displayed.
* When merging questions, the merged headline is carried over into the
  remaining question's keywords field (and thus contributes to search results)