Allow anonymous users to reuse a user account's email address.
The most common case here will be people who've created an account to submit
a question, and then don't bother logging in again after their session expires.
Making it easy for those people to use the site is a bigger priority than
preventing their identities from being falsely used to cast votes.  (We can,
in theory, eventually catch that sort of fraud via a user profile page and/or
IP address and other header inspection, if it becomes important.)