Cache temp values longer
When we compute things like the number of votes,
set its expiration in the cache to be much longer.
We want it to last at least through a full deploy,
however long that takes, because once the worker
is updated, it'll start putting its work into a
new cache "version" while all the unupgraded servers
will still be looking for it in the old cache version.
So that old cache version's entries need to continue
to be available as long as servers are looking for them.