Unbreaking appearance of social share buttons
The bad wrapping of social share buttons was caused by the text wrapping of the contained `a` element; this was confusing, because the parent element was (correctly) set up to not wrap its contents.

The appearance of the social links, on small screen widths, was pretty rough. This hopefully improves that state of affairs.