Add logic to ignore scripted votes using the same session id
plus update the tests, including a test utility to reset the session
on the test client.

We are treating requests with duplicate session ids the same as if
they were an anonymous user using an email that already voted on this
question -- don't create the vote, but don't respond with an error
either.