[OP-223] Fix for error changing password
This was a more general problem that some URLs were only valid
with a prefix, and in some contexts like changing a password,
no prefix was known.

These changes:

* Make any URL that was valid with a prefix also valid with no prefix
* Modify the Debate model to allow an empty prefix

It might be necessary to always have a debate configured with no
prefix for everything to work... hopefully we can live with that
for now?